Rename ctsm_pylib to just ctsm
Recommendation is to not have underscores in package names

Also, add 'make clean'
