+++ python/ctsm/joblauncher/job_launcher_factory.py
@@ -3,8 +3,8 @@
 
 from __future__ import print_function
 
-from ctsm_pylib.joblauncher.job_launcher_no_batch import JobLauncherNoBatch
-from ctsm_pylib.joblauncher.job_launcher_qsub import JobLauncherQsub
+from ctsm.joblauncher.job_launcher_no_batch import JobLauncherNoBatch
+from ctsm.joblauncher.job_launcher_qsub import JobLauncherQsub
 import logging
 
 logger = logging.getLogger(__name__)
