+++ python/ctsm/joblauncher/job_launcher_no_batch.py
@@ -1,6 +1,6 @@
 """A JobLauncher for systems where we can run a big job directly on the login node"""
 
-from ctsm_pylib.joblauncher.job_launcher_base import JobLauncherBase
+from ctsm.joblauncher.job_launcher_base import JobLauncherBase
 
 class JobLauncherNoBatch(JobLauncherBase):
