+++ python/ctsm/machine_utils.py
